In repli, shrink num timing values and add more than enough rho values.
Timing was already quite fine, and we have an exact posterior distribution at this resolution. We do not care about the maximum a posteriori but only the MLE estimator (weighted mean). The estimator produces results (meaned by probability value) which are far finer than the grid size. On the other hand, we want the concentration value rho to self-evidently cover all of the grid area of interest.