Final updates to chic violin.
Our estimate for the extremes is if the locus is overlapped by 5 of the rolling windows. E.g. if there is an extreme of the enrichment, then we will only show it if at least for the number of rolling windows that could overlap the monosome center location, all of them show at least this enrichment value.